Refactor useAlarmQuery to include revalidateOnFocus and dedupingInterval options
This commit updates the useAlarmQuery hook to include the revalidateOnFocus and dedupingInterval options in the useSWRInfinite function. By setting revalidateOnFocus to false and dedupingInterval to 1000 * 60 * 10, we prevent unnecessary revalidation of data when the page regains focus and reduce duplicate requests within a 10-minute interval. This optimization improves the performance and efficiency of data fetching in the useAlarmQuery hook.